requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pydantic-settings
httpx
python-multipart
openai
